add proper bindings for StreetViewPanorama's position
- Update examples to show how two-way binding should really be done,
i.e. there is a single `mapCenter` object that controls the
map and the street view.
The position of the map / panorama is synchronized only when it
has been explicitly changed by the user (@dragend, @zoom_changed),
but not during panning (because during panning, the "source of
truth" should be the user's mouse position.