Bug Fix in ReadME documentation:
In order for this plugin to install all components automatically the
'installComponents' attribute must be explicitly set to true. This was
not clear in the documentation because the necessary line to correctly
set this attribute in the example was commented out, implying that the
attribute did not need to be included.